Reviewer comments.
